refactor(ui): Convert axis limit controls in `NodeGraphWindow` from `double` to `float` and update UI interactions

+++ dynamic-neural-field-composer/include/user_interface/node_graph_window.h
@@ -62,7 +62,7 @@ namespace dnf_composer::user_interface
 		ImVec2 initialPos     = {};
 		ImVec2 size           = { 380.0f, 300.0f };
 		bool   autoFit        = true;
-		double xMin = 0, xMax = 100, yMin = -20, yMax = 20;
+		float xMin = 0.f, xMax = 100.f, yMin = -20.f, yMax = 20.f;
 		float  lineThickness  = 2.5f;
 		char   xLabel[64]     = "Spatial dimension";
 		char   yLabel[64]     = "Amplitude";
